pyproject.toml
[project]
name = "project-run"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "boto3==1.37.37",
    "django==5.2",
    "django-debug-toolbar>=5.2.0",
    "django-filter==25.1",
    "django-storages==1.14.6",
    "djangorestframework==3.16.0",
    "geopy>=2.4.1",
    "mypy>=1.17.0",
    "openpyxl>=3.1.5",
    "pre-commit>=4.2.0",
    "psycopg2-binary==2.9.10",
    "ruff>=0.12.5",
]

[tool.mypy]
strict = false
ignore_missing_imports = true

[tool.ruff]
required-version = '>=0.12.5'
src = ['project_run']
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".venv",
]
line-length = 88
indent-width = 4
target-version = "py313"
lint.fixable = ["ALL"]
lint.select = [
    "DJ", # flake8-django
    "ANN", # flake8-annotations
    "I", # isort
    "F", # Pyflakes
    "COM", # flake8-commas
    "Q", # flake8-quotes
    "S", # flake8-bandit
    "BLE", # flake8-blind-except
    "FBT", # flake8-boolean-trap
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "T10", # flake8-debugger
    "EM", # flake8-errmsg
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "LOG", # flake8-logging
    "G", # flake8-logging-format
    "INP", # flake8-no-pep420
    "PIE", # flake8-pie
    "T20", # flake8-print
    "PYI", # flake8-pyi
    "RSE", # flake8-raise
    "RET", # flake8-return
    "SLF", # flake8-self
    "SIM", # flake8-simplify
    "TC", # flake8-type-checking
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
    "FLY", # flynt
    "C90", # mccabe
    "N", # pep8-naming
    "PERF", # Perflint
    "E", # Error
    "W", # Warning
    "PGH", # pygrep-hooks
    "PLC", # Convention
    "PLE", # Pylint Error
    "PLR", # Pylint Refactor
    "PLW", # Pylint Warning
    "UP", # pyupgrade
    "FURB", # refurb
    "RUF", # Ruff-specific rules
    "TRY", # tryceratops

]

lint.ignore = [
    "S311", # suspicious-non-cryptographic-random-usage
    "TC006", # runtime-cast-value
    "S101", # S101	assert	Use of assert detected
    "RUF001", # ambiguous-unicode-character-string (RUF001)
    "RUF002", # ambiguous-unicode-character-docstring (RUF002)
    "RUF003", # ambiguous-unicode-character-comment (RUF003)
    "RUF005",
    "F403",
    "F405",
    "S105",
    "EM101",
    "TRY003",
    "PLC0415",
    "PLR2004",
]

[tool.ruff.lint.extend-per-file-ignores]
"stuff.py" = [
    "T20", # flake8-print (T20)
]
